feat: Finish first round of integration
