--- src/sphinxnotes/any/errors.py
@@ -1,13 +1,0 @@
-"""
-sphinxnotes.any.errors
-~~~~~~~~~~~~~~~~~~~~~~
-
-:copyright: Copyright 2021 Shengyu Zhang
-:license: BSD, see LICENSE for details.
-"""
-
-from sphinx.errors import ExtensionError
-
-
-class AnyExtensionError(ExtensionError):
-    pass
